Move update request tests
