+++ apps/backend/src/tests/neighborhood_api.test.ts
@@ -45,10 +45,6 @@ const loginUser = async (loginData: LoginData): Promise<Response> => {
   return loginResponse;
 };
 
-beforeAll(async () => {
-  await testHelpers.removeAllData();
-});
-
 describe('Tests for getting all neighborhoods: GET /neighborhoods', () => {
   beforeAll(async () => {
     await seed();
@@ -906,155 +902,6 @@ describe('Tests for creating a new request at POST /requests', () => {
     expect(numberOfFinalRequestsAssociatedWithNeighborhood)
       .toEqual(numInitialRequestsAssociatedWithNeighborhood + 1);
     expect(neighborhoodsRequestTitlesAfterCreation).toContain('foofoo');
-  });
-});
-
-describe('Tests for updating a request: PUT /neighborhoods/:nId/requests/:rId', () => {
-  let token: string;
-
-  beforeAll(async () => {
-    await seed();
-
-    const loginResponse: Response = await loginUser(MIKES_LOGIN_DATA);
-    token = loginResponse.body.token;
-  });
-
-  afterEach(async () => {
-    await seed();
-  });
-
-  test('Update a request\'s title, content and status fields', async () => {
-    const newData: UpdateRequestData = {
-      title: 'Test',
-      content: 'Test',
-      status: 'CLOSED',
-    };
-
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`)
-      .send(newData);
-
-    const updatedRequest: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.body).toEqual(updatedRequest.body);
-    expect(response.status).toEqual(200);
-  });
-
-  test('Empty input doesn\'t change anything on the server', async () => {
-    const requestBeforeUpdate: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`)
-      .send({});
-
-    const request: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(request.body).toEqual(requestBeforeUpdate.body);
-    expect(response.status).toEqual(200);
-  });
-
-  test('User cannot update request if they didn\'t create it', async () => {
-    const loginResponse: Response = await loginUser(BOBS_LOGIN_DATA);
-    const bobToken: string = loginResponse.body.token;
-
-    const newData: UpdateRequestData = { title: 'Test' };
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${bobToken}`)
-      .send(newData);
-
-    // Bob is admin of current neighborhood, therefore can fetch the request
-    const request: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${bobToken}`);
-
-    expect(response.status).toBe(401);
-    expect(request.body!.title).not.toBe(newData.title);
-  });
-
-  test('User cannot update request if they aren\'t logged in', async () => {
-    const newData: UpdateRequestData = { title: 'Test' };
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .send(newData);
-
-    const request: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.status).toBe(401);
-    expect(request.body?.title).not.toBe(newData.title);
-  });
-
-  test('Update with invalid properties raises an error', async () => {
-    const invalidData = { invalid: 'Test' };
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`)
-      .send(invalidData);
-
-    const request: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.status).toBe(400);
-    expect(request.body).not.toHaveProperty('invalid');
-  });
-
-  test('Update with invalid values raises an error', async () => {
-    const invalidData = {
-      title: null,
-      content: 2,
-      status: 'random string',
-    };
-
-    const requestBeforeUpdate: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`)
-      .send(invalidData);
-
-    const request: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.status).toBe(400);
-    expect(request?.body).toEqual(requestBeforeUpdate?.body);
-  });
-
-  test('Id, user_id, neighborhood_id and time_created props cannot be edited', async () => {
-    const data = {
-      id: 10,
-      user_id: BOBS_USER_ID,
-      time_created: new Date(),
-    };
-
-    const requestBeforeUpdate: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    const response: Response = await api
-      .put(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`)
-      .send(data);
-
-    const request: Response = await api
-      .get(`/api/neighborhoods/${BOBS_NHOOD_ID}/requests/${MIKES_REQUEST_ID}`)
-      .set('Authorization', `Bearer ${token}`);
-
-    expect(response.status).toBe(400);
-    expect(request?.body).toEqual(requestBeforeUpdate?.body);
   });
 });
